+++ phoss-smp-webapp-sql/src/test/java/com/helger/phoss/smp/SPITest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.helger.peppol.smpserver;
+package com.helger.phoss.smp;
 
 import org.junit.Test;
